pbtogcs-cf/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.42.3
